Add page navigation to app
