+++ package.json
@@ -10,12 +10,16 @@
   },
   "dependencies": {
     "@react-native-community/checkbox": "^0.5.14",
+    "@react-navigation/native": "^6.1.1",
+    "@react-navigation/native-stack": "^6.9.7",
     "expo": "~47.0.9",
+    "expo-checkbox": "~2.2.2",
+    "expo-linear-gradient": "~12.0.1",
     "expo-status-bar": "~1.4.2",
     "react": "18.1.0",
     "react-native": "0.70.5",
-    "expo-checkbox": "~2.2.2",
-    "expo-linear-gradient": "~12.0.1"
+    "react-native-screens": "~3.18.0",
+    "react-native-safe-area-context": "4.4.1"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9"
